+++ src/app/types/category.ts
@@ -1,0 +1,5 @@
+import  {  Document, Types } from 'mongoose';
+export interface ICategory extends Document {
+    name: string;
+    recipes: Types.ObjectId[]; 
+  }
