+++ src/app/types/recipes.ts
@@ -1,0 +1,9 @@
+
+import { Document, Types } from 'mongoose';
+  export interface IRecipe extends Document {
+    image: string;
+    name: string;
+    category: Types.ObjectId;
+    ingredients: string[];
+    favorite: boolean;
+  }
